requirements.txt
fastapi
uvicorn[standard]
openai
numexpr

google-generativeai
portalocker
google-cloud-firestore
google-cloud-aiplatform
requests
httpx
transformers
